# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AuthResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.BuyDecorationRequest:
    properties:
      decoration:
        type: string
    required:
    - decoration
    type: object
  models.CreateTaskRequest:
    properties:
      description:
        type: string
      reward:
        minimum: 1
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  models.Decoration:
    properties:
      createdAt:
        type: string
      decoration:
        type: string
      userId:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.Pet:
    properties:
      exp:
        type: integer
      happiness:
        type: integer
      hunger:
        type: integer
      level:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  models.SyncRequest:
    properties:
      lastSyncAt:
        type: string
    type: object
  models.SyncResponse:
    properties:
      decorations:
        items:
          $ref: '#/definitions/models.Decoration'
        type: array
      pet:
        $ref: '#/definitions/models.Pet'
      syncedAt:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Task:
    properties:
      completed:
        type: boolean
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      reward:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      gold:
        type: integer
      id:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@guildquest.io
    name: API Support
    url: https://guildquest.io/support
  description: Production-ready gamified task management system with virtual pet
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://guildquest.io/terms
  title: GuildQuest API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /decorations:
    get:
      description: Get all decorations owned by user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Decoration'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get owned decorations
      tags:
      - decorations
  /decorations/buy:
    post:
      consumes:
      - application/json
      description: Purchase a decoration (costs 50 gold)
      parameters:
      - description: Decoration to buy
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BuyDecorationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buy decoration
      tags:
      - decorations
  /invite/{token}/qr:
    get:
      parameters:
      - description: Invite token
        in: path
        name: token
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get invite QR
      tags:
      - invite
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /pet:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pet'
      security:
      - BearerAuth: []
      summary: Get pet
      tags:
      - pet
  /sync:
    post:
      consumes:
      - application/json
      description: Sync all user data (PWA offline support)
      parameters:
      - description: Last sync timestamp
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SyncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SyncResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Sync data
      tags:
      - sync
  /tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
      security:
      - BearerAuth: []
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Task
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - tasks
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"